Updated project
 - SMStoreDidStartSyncingNotification now works for those who need it.
 - Updated project for Xcode 7.2
 - Fixed one error and one warning
 - Made iOS example tabular, now it has a new tab Settings which
explains how to store setting values in CoreData + back them up/sync
them with CloudKit. I also added a simple mechanism to sync process for
avoiding multiple records with same identifier, which in this case, is
a name of setting. Mechanism uses current date/time while saving
setting the first time. When data is synced and if more than one
records is found, the oldest one (cloud one) will be the one that is
stored and others will be removed.

It should be pretty simple to use this for other kind of entities too.